move testing helpers to tests/utils
- move testing.py to tests/mock_data.py
- update test imports

Did not move to tests/utils.py to avoid conflict with existing utils/ repo.

Fix #1564
